app/requirements.txt
numpy
Pillow
networkx
nltk
opencv-python
reportlab
ipywidgets
matplotlib
requests
BeautifulSoup4
sympy
scipy
fonttools